% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3views.ipcaps.R
\name{plot3views.ipcaps}
\alias{plot3views.ipcaps}
\title{Plot 3 views from the result of IPCAPS}
\usage{
plot3views.ipcaps(ipcaps_output)
}
\arguments{
\item{ipcaps_output}{the returned object from the \code{\link{ipcaps2}} function.}
}
\value{
\code{NULL}
}
\description{
Plot the first three components of PCA using in 3 views.
}
\details{
This function generates a plot in the standard R graphic. Therefore,
it can be saved to a file using png(), pdf(),  jpeg(), and tiff().
}
\examples{

# Importantly, bed file, bim file, and fam file are required
# Use the example files embedded in the package

BED.file <- system.file("extdata","ipcaps_example.bed",package="IPCAPS2")
LABEL.file <- system.file("extdata","ipcaps_example_individuals.txt.gz",package="IPCAPS2")

my.cluster <- ipcaps2(bed=BED.file,
                      label.file=LABEL.file,
                      lab.col=2,
                      out=tempdir(),
                      silence=TRUE,
                      no.rep=1)

# Show a plot
plot3views.ipcaps(my.cluster)
}
