% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw.data}
\alias{raw.data}
\title{Synthetic dataset containing single nucleotide polymorphisms (SNP)}
\format{
A matrix with 2,000 columns and 103 rows
}
\usage{
data(ipcaps_example)
}
\description{
The raw.data is the simulated dataset which consists of 2,000 independent SNPs
and 103 individuals belonging to one of 2 populations (50 individuals
each) and 3 outlying individuals. The matrix \code{raw.data} contains the number 0, 1,
and 2 representing SNP in additive coding.
}
\references{
Balding, D.J., and Nichols, R.A. (1995). A method for quantifying
differentiation between populations at multi-allelic loci and its
implications for investigating identity and paternity. Genetica 96, 3-12.
}
\seealso{
\code{\link{label}} and \code{\link{PC}}
}
\keyword{raw.data}
